requirements.txt
click==8.1.7
pytest==8.3.2
python-dotenv==1.0.1